[MOVE] GetHouseholdPort application.port.in 으로 이동
