+++ src/main/java/com/connect/accountApp/domain/household/application/service/GetHouseholdSettlementNowSettlement.java
@@ -7,9 +7,9 @@ import com.connect.accountApp.domain.expense.application.port.in.command.TotalEx
 import com.connect.accountApp.domain.expense.application.port.out.GetHouseholdTotalExpensePort;
 import com.connect.accountApp.domain.expense.application.port.out.GetTotalExpensePort;
 import com.connect.accountApp.domain.expense.application.port.out.command.TotalExpenseCommand;
-import com.connect.accountApp.domain.household.adapter.out.persistence.GetHouseholdPort;
 import com.connect.accountApp.domain.household.application.port.in.GetHouseholdSettlementUseCase;
 import com.connect.accountApp.domain.household.application.port.in.command.HouseholdNowCommand;
+import com.connect.accountApp.domain.household.application.port.out.GetHouseholdPort;
 import com.connect.accountApp.domain.household.domain.model.Household;
 import com.connect.accountApp.domain.titleuser.application.port.out.FindUserTitlePort;
 import com.connect.accountApp.domain.titleuser.application.port.out.command.UserTitleCommand;
