+++ src/main/java/com/connect/accountApp/domain/user/application/service/GetSendMoneyService.java
@@ -2,7 +2,7 @@ package com.connect.accountApp.domain.user.application.service;
 
 import com.connect.accountApp.domain.expense.application.port.out.GetHouseholdTotalExpensePort;
 import com.connect.accountApp.domain.expense.application.port.out.command.TotalExpenseCommand;
-import com.connect.accountApp.domain.household.adapter.out.persistence.GetHouseholdPort;
+import com.connect.accountApp.domain.household.application.port.out.GetHouseholdPort;
 import com.connect.accountApp.domain.household.domain.model.Household;
 import com.connect.accountApp.domain.user.application.port.in.GetSendMoneyUseCase;
 import com.connect.accountApp.domain.user.application.port.in.command.SendMoneyCommand;
